DisplayLastFive and minor fixes
Added the view to display the result of the last five runs in a cool
way.
Fixed mixed file names displayLastFive and displayStreak.
Fixed a query error in the function getLastFive in DRDB.php.
Moved .success and .failed to main.css instead of addTodayRun.css.
Added a condition to display addTodayRun wider.
Added instructions to display all the .parts more coherently.